Recalculate variable value on each use and allow redefinition
Instead of storing the current value of a variable, recalculate it
on each use. This helps with the case when a variable references
another variable that suddenly changes. On the down-side this
adds a risk that recursion will create an never-ending loop.

Also add support for redefining variables by storing with the same
name as is already available.